[DEV-2] Refactor task_manager into package
