+++ .tasks/DEV/DEV-2.json
@@ -2,7 +2,7 @@
   "id": "DEV-2",
   "title": "Refactor task_manager.py code structure",
   "description": "Split the monolithic task_manager.py file into smaller, more maintainable modules. Separate concerns like TaskManager class, CLI parsing, TUI interface, and utility functions into different files.",
-  "status": "todo",
+  "status": "in_progress",
   "comments": [
     {
       "id": 1,
@@ -18,8 +18,28 @@
       "id": 3,
       "text": "Current file is 611 lines - too large for a single module. Key areas to extract: TaskManager class (~300 lines), CLI parsing (~200 lines), TUI interface (~100 lines), utility functions (~20 lines)",
       "created_at": 1748515941.752537
+    },
+    {
+      "id": 4,
+      "text": "Started work on refactoring into modules",
+      "created_at": 1748516275.5507393
+    },
+    {
+      "id": 5,
+      "text": "Split monolithic script into package structure",
+      "created_at": 1748516411.547792
+    },
+    {
+      "id": 6,
+      "text": "Implemented core, utils, tui, cli modules",
+      "created_at": 1748516453.7397463
+    },
+    {
+      "id": 7,
+      "text": "All tests passing",
+      "created_at": 1748516472.1735702
     }
   ],
   "created_at": 1748515927.1541288,
-  "updated_at": 1748515941.75255
+  "updated_at": 1748516472.1736097
 }
