+++ task-manager/task_manager/__main__.py
@@ -1,0 +1,5 @@
+from .cli import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
+
