+++ task-manager/task_manager/tui.py
@@ -1,0 +1,102 @@
+import sys
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .core import TaskManager
+
+def launch_tui(tm: "TaskManager") -> None:
+    """Launch the Textual TUI for the task manager."""
+    if TYPE_CHECKING:  # pragma: no cover - type hints only
+        from textual.app import App, ComposeResult  # type: ignore
+        from textual.widgets import Header, Footer, Button, Static, Input, DataTable  # type: ignore
+        from textual.containers import Vertical  # type: ignore
+
+    try:
+        from textual.app import App, ComposeResult  # type: ignore
+        from textual.widgets import Header, Footer, Button, Static, Input, DataTable  # type: ignore
+        from textual.containers import Vertical  # type: ignore
+    except Exception:  # pragma: no cover - optional dependency
+        print(
+            "Textual is required for the UI. Install with 'pip install textual'",
+            file=sys.stderr,
+        )
+        return
+
+    class TMApp(App):
+        TITLE = "Task Manager"
+        BINDINGS = [("q", "quit", "Quit")]
+
+        def __init__(self, manager: "TaskManager") -> None:
+            super().__init__()
+            self.manager = manager
+            self.body: Vertical | None = None
+
+        def compose(self) -> ComposeResult:
+            yield Header(show_clock=True)
+            yield Footer()
+            self.body = Vertical()
+            yield self.body
+
+        def on_mount(self) -> None:
+            """Called when the app is mounted and ready."""
+            self.show_main()
+
+        def show_main(self) -> None:
+            assert self.body
+            self.body.remove_children()
+            self.body.mount(Static("Task Manager", classes="title"))
+            self.body.mount(Button("Queues", id="queues"))
+            self.body.mount(Button("Tasks", id="tasks"))
+            self.body.mount(Button("Quit", id="quit"))
+
+        def show_queues(self) -> None:
+            assert self.body
+            self.body.remove_children()
+            table: DataTable = DataTable()
+            table.add_columns("Name", "Title", "Description")
+            for q in self.manager.queue_list():
+                table.add_row(q["name"], q["title"], q["description"])
+            self.body.mount(table)
+            self.body.mount(Button("Add Queue", id="queue_add"))
+            self.body.mount(Button("Back", id="main"))
+
+        def show_tasks(self) -> None:
+            assert self.body
+            self.body.remove_children()
+            table: DataTable = DataTable()
+            table.add_columns("ID", "Title", "Status")
+            for t in self.manager.task_list():
+                table.add_row(t["id"], t["title"], t["status"])
+            self.body.mount(table)
+            self.body.mount(Button("Back", id="main"))
+
+        def on_button_pressed(self, event: Button.Pressed) -> None:  # pragma: no cover - UI callbacks
+            button_id = event.button.id
+            if button_id == "quit":
+                self.exit()
+            elif button_id == "queues":
+                self.show_queues()
+            elif button_id == "tasks":
+                self.show_tasks()
+            elif button_id == "main":
+                self.show_main()
+            elif button_id == "queue_add":
+                assert self.body
+                self.body.remove_children()
+                name_in = Input(placeholder="Queue name", id="q_name")
+                title_in = Input(placeholder="Queue title", id="q_title")
+                desc_in = Input(placeholder="Description", id="q_desc")
+                self.body.mount(name_in)
+                self.body.mount(title_in)
+                self.body.mount(desc_in)
+                self.body.mount(Button("Create", id="create_queue"))
+                self.body.mount(Button("Back", id="queues"))
+            elif button_id == "create_queue":
+                name = self.query_one("#q_name", Input).value
+                title = self.query_one("#q_title", Input).value
+                desc = self.query_one("#q_desc", Input).value
+                self.manager.queue_add(name, title, desc)
+                self.show_queues()
+
+    TMApp(tm).run()
+
